+++ packages/instrumentation-dataloader/test/dataloader.test.ts
@@ -1,17 +1,6 @@
 /*
  * Copyright The OpenTelemetry Authors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
+ * SPDX-License-Identifier: Apache-2.0
  */
 
 import {
@@ -92,7 +81,7 @@ describe('DataloaderInstrumentation', () => {
       assert.strictEqual(batchSpan.name, 'dataloader.batch');
       assert.strictEqual(batchSpan.kind, SpanKind.INTERNAL);
       assert.deepStrictEqual(batchSpan.links, [
-        { context: loadSpan.spanContext(), attributes: {} },
+        { context: loadSpan.spanContext() },
       ]);
     });
 
@@ -206,7 +195,7 @@ describe('DataloaderInstrumentation', () => {
       assert.strictEqual(batchSpan.name, 'dataloader.batch');
       assert.strictEqual(batchSpan.kind, SpanKind.INTERNAL);
       assert.deepStrictEqual(batchSpan.links, [
-        { context: loadSpan.spanContext(), attributes: {} },
+        { context: loadSpan.spanContext() },
       ]);
 
       assert.strictEqual(loadManySpan.name, 'dataloader.loadMany');