+++ packages/resource-detector-azure/test/detectors/AzureFunctionsDetector.test.ts
@@ -1,22 +1,14 @@
 /*
  * Copyright The OpenTelemetry Authors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
+ * SPDX-License-Identifier: Apache-2.0
  */
 
 import * as assert from 'assert';
-import { azureFunctionsDetector } from '../../src/detectors/AzureFunctionsDetector';
-import { azureAppServiceDetector } from '../../src/detectors/AzureAppServiceDetector';
+import {
+  azureAppServiceDetector,
+  azureContainerAppsDetector,
+  azureFunctionsDetector,
+} from '../../src';
 import { ATTR_SERVICE_NAME } from '@opentelemetry/semantic-conventions';
 import {
   ATTR_CLOUD_PLATFORM,
@@ -50,7 +42,11 @@ describe('AzureFunctionsDetector', () => {
     process.env.WEBSITE_RESOURCE_GROUP = 'test-resource-group';
 
     const resource = detectResources({
-      detectors: [azureFunctionsDetector, azureAppServiceDetector],
+      detectors: [
+        azureFunctionsDetector,
+        azureAppServiceDetector,
+        azureContainerAppsDetector,
+      ],
     });
     assert.ok(resource);
     const attributes = resource.attributes;
@@ -86,7 +82,11 @@ describe('AzureFunctionsDetector', () => {
 
     const expectedWebsiteOwnerName = 'test-owner-name';
     const resource = detectResources({
-      detectors: [azureFunctionsDetector, azureAppServiceDetector],
+      detectors: [
+        azureFunctionsDetector,
+        azureAppServiceDetector,
+        azureContainerAppsDetector,
+      ],
     });
     assert.ok(resource);
     const attributes = resource.attributes;
@@ -108,7 +108,11 @@ it('should detect azure functions if websiteSku is defined as FlexConsumption', 
   process.env.WEBSITE_RESOURCE_GROUP = 'test-resource-group';
 
   const resource = detectResources({
-    detectors: [azureFunctionsDetector, azureAppServiceDetector],
+    detectors: [
+      azureFunctionsDetector,
+      azureAppServiceDetector,
+      azureContainerAppsDetector,
+    ],
   });
   assert.ok(resource);
   const attributes = resource.attributes;