+++ packages/instrumentation-knex/CHANGELOG.md
@@ -1,6 +1,25 @@
 <!-- markdownlint-disable MD007 MD034 -->
 # Changelog
 
+## [0.59.0](https://github.com/open-telemetry/opentelemetry-js-contrib/compare/instrumentation-knex-v0.58.0...instrumentation-knex-v0.59.0) (2026-04-17)
+
+
+### Features
+
+* **deps:** update deps matching '@opentelemetry/*' ([#3479](https://github.com/open-telemetry/opentelemetry-js-contrib/issues/3479)) ([8891261](https://github.com/open-telemetry/opentelemetry-js-contrib/commit/8891261cb590efcb661bd9f8afec4d1adf885ad8))
+
+## [0.58.0](https://github.com/open-telemetry/opentelemetry-js-contrib/compare/instrumentation-knex-v0.57.0...instrumentation-knex-v0.58.0) (2026-03-25)
+
+
+### Features
+
+* **deps:** update deps matching '@opentelemetry/*' ([#3450](https://github.com/open-telemetry/opentelemetry-js-contrib/issues/3450)) ([c8df394](https://github.com/open-telemetry/opentelemetry-js-contrib/commit/c8df394f02d68ae48a79a50258682c09dac13b8b))
+
+
+### Bug Fixes
+
+* **instrumentation-knex:** parse connectionString for db name, host and port ([#3424](https://github.com/open-telemetry/opentelemetry-js-contrib/issues/3424)) ([3740fd8](https://github.com/open-telemetry/opentelemetry-js-contrib/commit/3740fd89f8252906b4470146a9a5bfd4827fee34))
+
 ## [0.57.0](https://github.com/open-telemetry/opentelemetry-js-contrib/compare/instrumentation-knex-v0.56.0...instrumentation-knex-v0.57.0) (2026-03-04)